Return 4 is battery is charged
On OS X, if AC power is removed, and restored with the battery at 95-99%,
the system will keep the battery at that level. Liquid prompt would show the
green charging, even if the battery will never get to 100%.

Since this was done on OS X, with a pmset->acpi wrapper, I don't know if
acpi on linux capitalises Charged.

BSDers are on their own.
OS X likes to keep the battery between 99-95%